requirements.txt
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
openai>=1.0.0
tavily-python>=0.3.0
google-search-results>=2.4.2
deepagents>=0.1.0
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.0.0

# dev / CI
ruff>=0.4.0
pytest>=8.0.0
